requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.8
pdfminer.six==20221105
werkzeug==3.0.1
python-dateutil==2.8.2
pandas==2.1.4
